Moving FluentUIHostingController to its own file.
